Light_Manager/desktopV2/ΩδΓφΩí_ƒΘΩ∞δº½_v3/requirements.txt
# متطلبات منظومة المهندس v3
mysql-connector-python>=8.0.0
PyMySQL>=1.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
openpyxl>=3.0.0
reportlab>=3.6.0
Pillow>=8.0.0
python-dateutil>=2.8.0
PySide6>=6.6